usePrintView: move to store/app/hooks
