+++ webapp/components/Assessment/DataTable/Chart/chartWrapper.tsx
@@ -2,7 +2,8 @@ import React, { useEffect, useRef, useState } from 'react'
 
 import { elementOffset } from '@webapp/utils/domUtils'
 
-import { useOnResize, usePrintView } from '@webapp/components/hooks'
+import { useOnResize } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
 
 import ChartContainer from './chartContainer'
