+++ webapp/components/Assessment/DataTable/DataTable.tsx
@@ -5,7 +5,8 @@ import { useSelector } from 'react-redux'
 import { AssessmentType, TableData } from '@core/assessment'
 import { Objects } from '@core/utils'
 import { TableSpec } from '@webapp/sectionSpec'
-import { useI18n, usePrintView } from '@webapp/components/hooks'
+import { useI18n } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
 
 import Table from './Table'
 import Chart from './Chart'
@@ -74,7 +75,7 @@ const DataTable: React.FC<Props> = (props) => {
         />
       )}
 
-      {printView && breakPointsColsPrint.length > 0 ? (
+      {printView && breakPointsColsPrint?.length > 0 ? (
         breakPointsColsPrint.map((breakPoint, idx) => {
           const rowsSliced = getRowsSliced(breakPointsColsPrint, idx, rows)
           return (
