+++ webapp/components/Assessment/DataTable/Table/CellOdpHeader/CellOdpHeader.tsx
@@ -3,7 +3,8 @@ import { Link } from 'react-router-dom'
 
 import { FRA, TableDatumODP } from '@core/assessment'
 import * as BasePaths from '@webapp/main/basePaths'
-import { useCountryIso, useI18n, usePrintView, useUserInfo } from '@webapp/components/hooks'
+import { useCountryIso, useI18n, useUserInfo } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
 
 import Icon from '@webapp/components/icon'
 import Tooltip from '@webapp/components/tooltip'
