+++ webapp/components/Assessment/FraPrint/FraPrint.tsx
@@ -4,11 +4,12 @@ import { useSelector } from 'react-redux'
 
 import { Assessment, FRA } from '@core/assessment'
 import * as CountryState from '@webapp/app/country/countryState'
-import { useCountryIso, useI18n, usePrintView } from '@webapp/components/hooks'
+import { useCountryIso, useI18n } from '@webapp/components/hooks'
 
 import SectionView from '@webapp/components/Assessment/SectionView'
 import Loading from '@webapp/components/loading'
 import ContactPersonsPrintView from '@webapp/sectionSpec/fra/contactPersons/contactPersonsPrintView'
+import { usePrintView } from '@webapp/store/app'
 import TableOfContent from './TableOfContent'
 import useFraPrintDataFetch from './useFraPrintDataFetch'
