+++ webapp/components/Assessment/SectionView/Descriptions/Descriptions.tsx
@@ -3,7 +3,8 @@ import { useSelector } from 'react-redux'
 
 import { Objects } from '@core/utils'
 import { DescriptionsSpec } from '@webapp/sectionSpec'
-import { useI18n, usePrintView } from '@webapp/components/hooks'
+import { useI18n } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
 
 import NationalDataDescriptions from './NationalDataDescriptions'
 import AnalysisDescriptions from './AnalysisDescriptions'
