+++ webapp/components/Assessment/SectionView/SectionView.tsx
@@ -4,7 +4,9 @@ import { useSelector } from 'react-redux'
 
 import { AssessmentType } from '@core/assessment'
 import { SectionSpecs } from '@webapp/sectionSpec'
-import { useI18n, usePrintView } from '@webapp/components/hooks'
+import { useI18n } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
+
 import * as FraState from '@webapp/app/assessment/fra/fraState'
 
 import DataTable from '@webapp/components/Assessment/DataTable'
