+++ webapp/components/Assessment/SectionView/Title/ExtentOfForest/ExtentOfForest.tsx
@@ -4,7 +4,8 @@ import { useDispatch, useSelector } from 'react-redux'
 import { Users } from '@core/auth'
 import * as ExtentOfForestState from '@webapp/sectionSpec/fra/extentOfForest/extentOfForestState'
 import { toggleOdps } from '@webapp/sectionSpec/fra/extentOfForest/actions'
-import { useI18n, usePrintView, useUserInfo } from '@webapp/components/hooks'
+import { useI18n, useUserInfo } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
 
 import OriginalDataPointsPrint from '@webapp/components/OriginalDataPoint/OriginalDataPointsPrint'
 import { Props } from '../props'
