+++ webapp/components/Assessment/SectionView/Title/ForestCharacteristics/ForestCharacteristics.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { useSelector } from 'react-redux'
 
 import * as ForestCharacteristicsState from '@webapp/sectionSpec/fra/forestCharacteristics/forestCharacteristicsState'
-import { useI18n, usePrintView } from '@webapp/components/hooks'
+import { useI18n } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
 
 import OriginalDataPointsPrint from '@webapp/components/OriginalDataPoint/OriginalDataPointsPrint'
 import { Props } from '../props'
