+++ webapp/components/OriginalDataPoint/ExtentOfForest/ExtentOfForest.tsx
@@ -3,7 +3,9 @@ import React from 'react'
 import { ODP, ODPs } from '@core/odp'
 import * as NumberFormat from '@common/numberFormat'
 import DefinitionLink from '@webapp/components/definitionLink'
-import { useI18n, usePrintView } from '@webapp/components/hooks'
+import { useI18n } from '@webapp/components/hooks'
+import { usePrintView } from '@webapp/store/app'
+
 import ExtentOfForestRow from './ExtentOfForestRow'
 
 type Props = {
