+++ webapp/hooks/index.ts
@@ -1,4 +1,3 @@
 export { default as useI18n } from './useI18n'
 export { default as useCountryIso } from './useCountryIso'
-export { default as usePrintView } from './usePrintView'
 export { default as useInitCountry } from './useInitCountry'
