--- webapp/hooks/usePrintView.ts
@@ -1,4 +1,0 @@
-import { useSelector } from 'react-redux'
-import * as AppState from '@webapp/store/app/state'
-
-export default () => useSelector((state) => [AppState.isPrintView(state), AppState.isPrintOnlyTablesView(state)])
