+++ webapp/store/app/index.ts
@@ -6,6 +6,7 @@ export {
   useGroupedRegions,
   useSecondaryGroupedRegions,
   useAppLoaded,
+  usePrintView,
 } from './hooks'
 
 export { default as AppReducer, AppActions } from './appSlice'
